gh-1 [wip] but making some good progress, need to resolve the toolbar issues
