+++ src/pymbe/api.py
@@ -1,6 +1,6 @@
 from .widget.client import SysML2ClientWidget
 from .widget.containment import ContainmentTree
+from .widget.diagram import SysML2LPGWidget
 from .widget.inspector import ElementInspector
-from .widget.graph import SysML2LPGWidget
 from .widget.ui import UI
 from .interpretation.interpretation import Instance
