+++ src/pymbe/widget/diagram/__init__.py
@@ -1,0 +1,1 @@
+from .widget import SysML2LPGWidget
